Split api model generators
