+++ app/src/test/scala/org/alephium/explorer/persistence/queries/BlockQueriesSpec.scala
@@ -24,6 +24,7 @@ import org.scalatest.time.{Minutes, Span}
 import slick.jdbc.PostgresProfile.api._
 
 import org.alephium.explorer.{AlephiumSpec, Generators}
+import org.alephium.explorer.GenApiModel.blockEntryHashGen
 import org.alephium.explorer.persistence.{DatabaseFixtureForEach, DBRunner}
 import org.alephium.explorer.persistence.schema._
