+++ app/src/test/scala/org/alephium/explorer/persistence/queries/TransactionQueriesSpec.scala
@@ -25,6 +25,8 @@ import org.scalatest.time.{Minutes, Span}
 import slick.jdbc.PostgresProfile.api._
 
 import org.alephium.explorer.{AlephiumSpec, Generators, Hash}
+import org.alephium.explorer.GenApiModel._
+import org.alephium.explorer.GenCoreUtil.timestampMaxValue
 import org.alephium.explorer.GenModel._
 import org.alephium.explorer.api.model._
 import org.alephium.explorer.persistence.{DatabaseFixtureForEach, DBRunner}
