+++ app/src/test/scala/org/alephium/explorer/service/BlockFlowSyncServiceSpec.scala
@@ -26,6 +26,8 @@ import org.scalatest.time.{Seconds, Span}
 
 import org.alephium.api.model.{ChainInfo, ChainParams, HashesAtHeight, SelfClique}
 import org.alephium.explorer.{AlephiumSpec, BlockHash, Generators, GroupSetting}
+import org.alephium.explorer.GenApiModel.chainIndexes
+import org.alephium.explorer.GenCoreUtil.timestampMaxValue
 import org.alephium.explorer.api.model._
 import org.alephium.explorer.cache.BlockCache
 import org.alephium.explorer.persistence.DatabaseFixtureForEach
