+++ app/src/test/scala/org/alephium/explorer/service/MempoolSyncServiceSpec.scala
@@ -27,6 +27,7 @@ import org.scalatest.time.{Minutes, Span}
 
 import org.alephium.api.model.{ChainInfo, ChainParams, HashesAtHeight, SelfClique}
 import org.alephium.explorer.{AlephiumSpec, Generators}
+import org.alephium.explorer.GenApiModel.utransactionGen
 import org.alephium.explorer.api.model.{BlockEntry, GroupIndex, Height, UnconfirmedTransaction}
 import org.alephium.explorer.persistence.DatabaseFixtureForEach
 import org.alephium.explorer.persistence.dao.UnconfirmedTxDao
