+++ app/src/test/scala/org/alephium/explorer/web/AddressServerSpec.scala
@@ -27,6 +27,7 @@ import slick.jdbc.PostgresProfile
 
 import org.alephium.api.ApiError
 import org.alephium.explorer.{AlephiumSpec, Generators, GroupSetting, Hash}
+import org.alephium.explorer.GenApiModel._
 import org.alephium.explorer.api.model._
 import org.alephium.explorer.cache.TransactionCache
 import org.alephium.explorer.persistence.DatabaseFixtureForEach
